+++ changelog.md
@@ -2,9 +2,8 @@
 
   * Add `thawByteArray` and `thawPrimArray`.
 
-  * Changed the `Show` instance of `ByteArray` so that all 8-bit words
-    are rendered as two digits. For example, display `0x0D` instead of `0xD`
-    for as an element.
+  * Changed the `Show` instance of `ByteArray`, so that all 8-bit words
+    are rendered as two digits. For example, display `0x0D` instead of `0xD`.
 
 ## Changes in version 0.7.1.0